refine draw boundary checks
